core/sfc-core/src/main/kotlin/com/amazonaws/sfc/transformations/Str.kt

// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
// SPDX-License-Identifier: MIT-0


package com.amazonaws.sfc.transformations

import com.amazonaws.sfc.config.ConfigurationClass
import com.google.gson.JsonObject

@ConfigurationClass
@TransformerOperator(["Str"])
class Str : TransformationImpl<Nothing>() {

    @TransformerMethod
    fun apply(target: Any?): String = target.toString()


    companion object {
        fun fromJson(o: JsonObject): TransformationOperator = TransformationOperatorNoOperand.fromJson<Str>(o)
        fun create() = Str()
    }
}